Copy search config from resources config singleton to search app
(Current effort is to make text extraction a function of search specifically)
