+++ solution/backend/content_search/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class FileSearchConfig(AppConfig):
     name = "content_search"
-    verbose_name = "postgres-based search for resources and regulation text"
+    verbose_name = "Content Search"
