--- solution/backend/resources/utils/__init__.py
@@ -1,2 +1,0 @@
-from .aws_utils import *  # noqa
-from .general_purpose import *  # noqa
