+++ solution/backend/resources/views/resources.py
@@ -21,7 +21,6 @@ from resources.models import (
     InternalFile,
     InternalLink,
     PublicLink,
-    ResourcesConfiguration,
     Subject,
 )
 from resources.serializers import (
@@ -38,7 +37,6 @@ from resources.serializers import (
 )
 from resources.utils import (
     CITATION_FILTER_PARAMETER,
-    call_text_extractor,
     get_citation_filter,
     string_to_bool,
 )
@@ -248,8 +246,10 @@ class FederalRegisterLinkViewSet(PublicResourceViewSet):
         sc = self.get_serializer(link, data=data, context={**self.get_serializer_context(), **{"created": created}})
         sc.is_valid(raise_exception=True)
         sc.save()
-        if ResourcesConfiguration.get_solo().auto_extract:
-            _, fail = call_text_extractor(request, FederalRegisterLink.objects.filter(pk=link.pk))
+        # if ResourcesConfiguration.get_solo().auto_extract:
+        #    _, fail = call_text_extractor(request, FederalRegisterLink.objects.filter(pk=link.pk))
+        if True:  # TODO FIX THIS
+            _, fail = (0, False)
             if fail:
                 logger.warning("Failed to extract text for Federal Register Link %i: %s", link.pk, fail[0]["reason"])
         return JsonResponse(sc.validated_data)
